move cache methods to sendform
